+++ udata_hydra/analysis/csv.py
@@ -71,13 +71,16 @@ async def analyse_csv(check_id: int = None, url: str = None, file_path: str = No
         log.debug("CSV_ANALYSIS_ENABLED turned off, skipping.")
         return
 
+    exceptions = config.LARGE_RESOURCES_EXCEPTIONS
+
     timer = Timer("analyse-csv")
     assert any(_ is not None for _ in (check_id, url))
     check = await get_check(check_id) if check_id is not None else {}
     url = check.get("url") or url
+    exception_file = str(check.get("resource_id", "")) in exceptions
 
     headers = json.loads(check.get("headers") or "{}")
-    tmp_file = open(file_path, "rb") if file_path else await download_resource(url, headers)
+    tmp_file = open(file_path, "rb") if file_path else await download_resource(url, headers, exception_file)
     table_name = hashlib.md5(url.encode("utf-8")).hexdigest()
     timer.mark("download-file")